1.firstScan 拆分完毕
